move reserve calendar to a component file and modify router
